# docs/openapi.yaml
openapi: 3.1.0
info:
  title: HIMS OPD API
  version: 1.0.0
  description: |
    REST API for the HIMS Hospital OPD Management System. All endpoints (except
    `/api/auth/login` and `/api/health`) require an authenticated session
    cookie issued by `/api/auth/login`.
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: hims_session
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
...
security:
  - sessionCookie: []
paths:
...
  /api/auth/login:
    post:
      summary: Sign in
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } }
      responses:
        '200': { description: Session cookie set }
        '401': { description: Invalid credentials }
...